Add files via upload
Uploading current files for page one of analysis. Original author Zachary Gvildys.

The .mml file is the model container, or the simulation box used in the actual page. The .cml file calls this file and they need to be in the same directory or file.

The .cml file is the file opened by java/molecular workbench.